Fix coverity for wifi wpa supplicant example
[Description]
Casting strlen(buf) from unsigned int to int without checking its value may result in
lost or misinterpreted data.

[Root Cause]
The size_t type has a larger range than the int type; using an int type variable to
store a size_t type variable may result in truncation.

[Fix]
use size_t to store.

[Testing]
compile ok

[Reviewers]
Gannicus Guo, Qingling Wu, Maochen Wang

[Port proposal]
non-general

+++ src/eap_server/eap_sim_db.c
@@ -733,7 +733,7 @@ static void process_eap_sim_db_event(struct eap_sim_db_data *data)
     void *mem;
     char *buf = NULL;
     char *pos, *cmd, *imsi;
-    int res;
+    size_t res;
 
 #ifdef __ZEPHYR__
     while (k_msgq_get(&eap_sim_db_event_queue, &mem, K_NO_WAIT) == 0)
